pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cert-llm-detector"
version = "1.0.0"
description = "TLS Certificate Phishing Detector with LLM Integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Security Research Team"}
]
keywords = ["security", "tls", "certificates", "phishing", "machine-learning", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "cryptography>=41.0.0",
    "pyOpenSSL>=23.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "google-generativeai>=0.3.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "python-Levenshtein>=0.21.0",
    "unidecode>=1.3.6",
]

[project.optional-dependencies]
ml = ["xgboost>=2.0.0", "imbalanced-learn>=0.11.0"]
rag = ["sentence-transformers>=2.2.0", "faiss-cpu>=1.7.4"]
viz = ["matplotlib>=3.7.0", "seaborn>=0.12.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
]
all = [
    "cert-llm-detector[ml,rag,viz,dev]"
]

[project.scripts]
cert-detector = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/cert-llm-detector"
Documentation = "https://github.com/yourusername/cert-llm-detector/blob/main/README.md"
Repository = "https://github.com/yourusername/cert-llm-detector"
Issues = "https://github.com/yourusername/cert-llm-detector/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]